chore (SG Mobile Data Usage Test Case) : Replace with Mockito
